Move API function into another script file
